+++ public/user_preferences.php
@@ -159,6 +159,7 @@ try {
   [$show_format_set, $show_format] = read_pref_bool($data, 'show_format');
   [$show_year_set, $show_year] = read_pref_bool($data, 'show_year');
   [$show_copy_count_set, $show_copy_count] = read_pref_bool($data, 'show_copy_count');
+  [$show_file_size_set, $show_file_size] = read_pref_bool($data, 'show_file_size');
   [$show_status_set, $show_status] = read_pref_bool($data, 'show_status');
   [$show_placement_set, $show_placement] = read_pref_bool($data, 'show_placement');
   [$show_isbn_set, $show_isbn] = read_pref_bool($data, 'show_isbn');
@@ -209,11 +210,11 @@ try {
   $sql = "INSERT INTO UserPreferences
           (user_id, logo_path, bg_color, fg_color, text_size, per_page,
           show_cover, show_subtitle, show_series, show_is_hungarian, show_publisher,
-          show_language, show_format, show_year, show_copy_count, show_status, show_placement, show_isbn, show_loaned_to,
+          show_language, show_format, show_year, show_copy_count, show_file_size, show_status, show_placement, show_isbn, show_loaned_to,
            show_loaned_date, show_subjects, show_notes, updated_at)
           VALUES (:uid, :logo_ins, :bg_ins, :fg_ins, :ts_ins, :per_ins,
            :show_cover_ins, :show_subtitle_ins, :show_series_ins, :show_is_hungarian_ins, :show_publisher_ins,
-           :show_language_ins, :show_format_ins, :show_year_ins, :show_copy_count_ins, :show_status_ins, :show_placement_ins, :show_isbn_ins, :show_loaned_to_ins,
+           :show_language_ins, :show_format_ins, :show_year_ins, :show_copy_count_ins, :show_file_size_ins, :show_status_ins, :show_placement_ins, :show_isbn_ins, :show_loaned_to_ins,
            :show_loaned_date_ins, :show_subjects_ins, :show_notes_ins, NOW())
           ON DUPLICATE KEY UPDATE
             logo_path = IF(:logo_set, :logo_upd, logo_path),
@@ -230,6 +231,7 @@ try {
             show_format = IF(:show_format_set, :show_format_upd, show_format),
             show_year = IF(:show_year_set, :show_year_upd, show_year),
             show_copy_count = IF(:show_copy_count_set, :show_copy_count_upd, show_copy_count),
+            show_file_size = IF(:show_file_size_set, :show_file_size_upd, show_file_size),
             show_status = IF(:show_status_set, :show_status_upd, show_status),
             show_placement = IF(:show_placement_set, :show_placement_upd, show_placement),
             show_isbn = IF(:show_isbn_set, :show_isbn_upd, show_isbn),
@@ -252,6 +254,7 @@ try {
     'show_format' => 0,
     'show_year' => 1,
     'show_copy_count' => 0,
+    'show_file_size' => 0,
     'show_status' => 1,
     'show_placement' => 1,
     'show_isbn' => 0,
@@ -278,6 +281,7 @@ try {
     ':show_format_ins' => $show_format ?? $defaults['show_format'],
     ':show_year_ins' => $show_year ?? $defaults['show_year'],
     ':show_copy_count_ins' => $show_copy_count ?? $defaults['show_copy_count'],
+    ':show_file_size_ins' => $show_file_size ?? $defaults['show_file_size'],
     ':show_status_ins' => $show_status ?? $defaults['show_status'],
     ':show_placement_ins' => $show_placement ?? $defaults['show_placement'],
     ':show_isbn_ins' => $show_isbn ?? $defaults['show_isbn'],
@@ -313,6 +317,8 @@ try {
     ':show_year_upd' => $show_year,
     ':show_copy_count_set' => $show_copy_count_set ? 1 : 0,
     ':show_copy_count_upd' => $show_copy_count,
+    ':show_file_size_set' => $show_file_size_set ? 1 : 0,
+    ':show_file_size_upd' => $show_file_size,
     ':show_status_set' => $show_status_set ? 1 : 0,
     ':show_status_upd' => $show_status,
     ':show_placement_set' => $show_placement_set ? 1 : 0,